rtl8723du: Replace NDIS_OID and NDIS_STATUS with uint, UCHAR with u8, USHORT with u16, and UINT and ULONG with u32

+++ core/rtw_mlme.c
@@ -863,7 +863,7 @@ bool rtw_update_scanned_network(_adapter *adapter, WLAN_BSSID_EX *target)
 {
 	_irqL irqL;
 	_list	*plist, *phead;
-	ULONG	bssid_ex_sz;
+	u32	bssid_ex_sz;
 	struct mlme_priv	*pmlmepriv = &(adapter->mlmepriv);
 	struct mlme_ext_priv	*pmlmeext = &(adapter->mlmeextpriv);
 	struct wifidirect_info *pwdinfo = &(adapter->wdinfo);
